modified code

+++ scripts/ssim.py
@@ -15,13 +15,13 @@ def _fspecial_gauss_1d(size, sigma):
     Returns:
         torch.Tensor: 1D kernel (1 x 1 x size)
     """
-    coords = torch.arange(size).to(dtype=torch.float)
+    coords = torch.arange( size ).to( dtype = torch.float )
     coords -= size // 2
 
-    g = torch.exp(-(coords ** 2) / (2 * sigma ** 2))
+    g = torch.exp( -(coords ** 2) / (2 * sigma ** 2) )
     g /= g.sum()
 
-    return g.unsqueeze(0).unsqueeze(0)
+    return g.unsqueeze( 0 ).unsqueeze( 0 )
 
 
 def gaussian_filter(input, win):
@@ -33,19 +33,19 @@ def gaussian_filter(input, win):
     Returns:
         torch.Tensor: blurred tensors
     """
-    assert all([ws == 1 for ws in win.shape[1:-1]]), win.shape
-    if len(input.shape) == 4:
+    assert all( [ws == 1 for ws in win.shape[1:-1]] ), win.shape
+    if len( input.shape ) == 4:
         conv = F.conv2d
-    elif len(input.shape) == 5:
+    elif len( input.shape ) == 5:
         conv = F.conv3d
     else:
-        raise NotImplementedError(input.shape)
+        raise NotImplementedError( input.shape )
 
     C = input.shape[1]
     out = input
-    for i, s in enumerate(input.shape[2:]):
+    for i, s in enumerate( input.shape[2:] ):
         if s >= win.shape[-1]:
-            out = conv(out, weight=win.transpose(2 + i, -1), stride=1, padding=0, groups=C)
+            out = conv( out, weight = win.transpose( 2 + i, -1 ), stride = 1, padding = 0, groups = C )
         else:
             warnings.warn(
                 f"Skipping Gaussian Smoothing at dimension 2+{i} for input: {input.shape} and win size: {win.shape[-1]}"
@@ -55,7 +55,6 @@ def gaussian_filter(input, win):
 
 
 def _ssim(X, Y, data_range, win, size_average=True, K=(0.01, 0.03)):
-
     r""" Calculate ssim index for X and Y
 
     Args:
@@ -75,37 +74,37 @@ def _ssim(X, Y, data_range, win, size_average=True, K=(0.01, 0.03)):
     C1 = (K1 * data_range) ** 2
     C2 = (K2 * data_range) ** 2
 
-    win = win.to(X.device, dtype=X.dtype)
+    win = win.to( X.device, dtype = X.dtype )
 
-    mu1 = gaussian_filter(X, win)
-    mu2 = gaussian_filter(Y, win)
+    mu1 = gaussian_filter( X, win )
+    mu2 = gaussian_filter( Y, win )
 
-    mu1_sq = mu1.pow(2)
-    mu2_sq = mu2.pow(2)
+    mu1_sq = mu1.pow( 2 )
+    mu2_sq = mu2.pow( 2 )
     mu1_mu2 = mu1 * mu2
 
-    sigma1_sq = compensation * (gaussian_filter(X * X, win) - mu1_sq)
-    sigma2_sq = compensation * (gaussian_filter(Y * Y, win) - mu2_sq)
-    sigma12 = compensation * (gaussian_filter(X * Y, win) - mu1_mu2)
+    sigma1_sq = compensation * (gaussian_filter( X * X, win ) - mu1_sq)
+    sigma2_sq = compensation * (gaussian_filter( Y * Y, win ) - mu2_sq)
+    sigma12 = compensation * (gaussian_filter( X * Y, win ) - mu1_mu2)
 
     cs_map = (2 * sigma12 + C2) / (sigma1_sq + sigma2_sq + C2)  # set alpha=beta=gamma=1
     ssim_map = ((2 * mu1_mu2 + C1) / (mu1_sq + mu2_sq + C1)) * cs_map
 
-    ssim_per_channel = torch.flatten(ssim_map, 2).mean(-1)
-    cs = torch.flatten(cs_map, 2).mean(-1)
+    ssim_per_channel = torch.flatten( ssim_map, 2 ).mean( -1 )
+    cs = torch.flatten( cs_map, 2 ).mean( -1 )
     return ssim_per_channel, cs
 
 
 def ssim(
-    X,
-    Y,
-    data_range=255,
-    size_average=True,
-    win_size=11,
-    win_sigma=1.5,
-    win=None,
-    K=(0.01, 0.03),
-    nonnegative_ssim=False,
+        X,
+        Y,
+        data_range=255,
+        size_average=True,
+        win_size=11,
+        win_sigma=1.5,
+        win=None,
+        K=(0.01, 0.03),
+        nonnegative_ssim=False,
 ):
     r""" interface of ssim
     Args:
@@ -123,42 +122,41 @@ def ssim(
         torch.Tensor: ssim results
     """
     if not X.shape == Y.shape:
-        raise ValueError("Input images should have the same dimensions.")
+        raise ValueError( "Input images should have the same dimensions." )
 
-    for d in range(len(X.shape) - 1, 1, -1):
-        X = X.squeeze(dim=d)
-        Y = Y.squeeze(dim=d)
+    for d in range( len( X.shape ) - 1, 1, -1 ):
+        X = X.squeeze( dim = d )
+        Y = Y.squeeze( dim = d )
 
-    if len(X.shape) not in (4, 5):
-        raise ValueError(f"Input images should be 4-d or 5-d tensors, but got {X.shape}")
+    if len( X.shape ) not in (4, 5):
+        raise ValueError( f"Input images should be 4-d or 5-d tensors, but got {X.shape}" )
 
     if not X.type() == Y.type():
-        raise ValueError("Input images should have the same dtype.")
+        raise ValueError( "Input images should have the same dtype." )
 
     if win is not None:  # set win_size
         win_size = win.shape[-1]
 
     if not (win_size % 2 == 1):
-        raise ValueError("Window size should be odd.")
+        raise ValueError( "Window size should be odd." )
 
     if win is None:
-        win = _fspecial_gauss_1d(win_size, win_sigma)
-        win = win.repeat([X.shape[1]] + [1] * (len(X.shape) - 1))
+        win = _fspecial_gauss_1d( win_size, win_sigma )
+        win = win.repeat( [X.shape[1]] + [1] * (len( X.shape ) - 1) )
 
-    ssim_per_channel, cs = _ssim(X, Y, data_range=data_range, win=win, size_average=False, K=K)
+    ssim_per_channel, cs = _ssim( X, Y, data_range = data_range, win = win, size_average = False, K = K )
     if nonnegative_ssim:
-        ssim_per_channel = torch.relu(ssim_per_channel)
+        ssim_per_channel = torch.relu( ssim_per_channel )
 
     if size_average:
         return ssim_per_channel.mean()
     else:
-        return ssim_per_channel.mean(1)
+        return ssim_per_channel.mean( 1 )
 
 
 def ms_ssim(
-    X, Y, data_range=255, size_average=True, win_size=11, win_sigma=1.5, win=None, weights=None, K=(0.01, 0.03)
+        X, Y, data_range=255, size_average=True, win_size=11, win_sigma=1.5, win=None, weights=None, K=(0.01, 0.03)
 ):
-
     r""" interface of ms-ssim
     Args:
         X (torch.Tensor): a batch of images, (N,C,[T,]H,W)
@@ -174,73 +172,73 @@ def ms_ssim(
         torch.Tensor: ms-ssim results
     """
     if not X.shape == Y.shape:
-        raise ValueError("Input images should have the same dimensions.")
+        raise ValueError( "Input images should have the same dimensions." )
 
-    for d in range(len(X.shape) - 1, 1, -1):
-        X = X.squeeze(dim=d)
-        Y = Y.squeeze(dim=d)
+    for d in range( len( X.shape ) - 1, 1, -1 ):
+        X = X.squeeze( dim = d )
+        Y = Y.squeeze( dim = d )
 
     if not X.type() == Y.type():
-        raise ValueError("Input images should have the same dtype.")
+        raise ValueError( "Input images should have the same dtype." )
 
-    if len(X.shape) == 4:
+    if len( X.shape ) == 4:
         avg_pool = F.avg_pool2d
-    elif len(X.shape) == 5:
+    elif len( X.shape ) == 5:
         avg_pool = F.avg_pool3d
     else:
-        raise ValueError(f"Input images should be 4-d or 5-d tensors, but got {X.shape}")
+        raise ValueError( f"Input images should be 4-d or 5-d tensors, but got {X.shape}" )
 
     if win is not None:  # set win_size
         win_size = win.shape[-1]
 
     if not (win_size % 2 == 1):
-        raise ValueError("Window size should be odd.")
+        raise ValueError( "Window size should be odd." )
 
-    smaller_side = min(X.shape[-2:])
+    smaller_side = min( X.shape[-2:] )
     assert smaller_side > (win_size - 1) * (
-        2 ** 4
+            2 ** 4
     ), "Image size should be larger than %d due to the 4 downsamplings in ms-ssim" % ((win_size - 1) * (2 ** 4))
 
     if weights is None:
         weights = [0.0448, 0.2856, 0.3001, 0.2363, 0.1333]
-    weights = torch.FloatTensor(weights).to(X.device, dtype=X.dtype)
+    weights = torch.FloatTensor( weights ).to( X.device, dtype = X.dtype )
 
     if win is None:
-        win = _fspecial_gauss_1d(win_size, win_sigma)
-        win = win.repeat([X.shape[1]] + [1] * (len(X.shape) - 1))
+        win = _fspecial_gauss_1d( win_size, win_sigma )
+        win = win.repeat( [X.shape[1]] + [1] * (len( X.shape ) - 1) )
 
     levels = weights.shape[0]
     mcs = []
-    for i in range(levels):
-        ssim_per_channel, cs = _ssim(X, Y, win=win, data_range=data_range, size_average=False, K=K)
+    for i in range( levels ):
+        ssim_per_channel, cs = _ssim( X, Y, win = win, data_range = data_range, size_average = False, K = K )
 
         if i < levels - 1:
-            mcs.append(torch.relu(cs))
+            mcs.append( torch.relu( cs ) )
             padding = [s % 2 for s in X.shape[2:]]
-            X = avg_pool(X, kernel_size=2, padding=padding)
-            Y = avg_pool(Y, kernel_size=2, padding=padding)
+            X = avg_pool( X, kernel_size = 2, padding = padding )
+            Y = avg_pool( Y, kernel_size = 2, padding = padding )
 
-    ssim_per_channel = torch.relu(ssim_per_channel)  # (batch, channel)
-    mcs_and_ssim = torch.stack(mcs + [ssim_per_channel], dim=0)  # (level, batch, channel)
-    ms_ssim_val = torch.prod(mcs_and_ssim ** weights.view(-1, 1, 1), dim=0)
+    ssim_per_channel = torch.relu( ssim_per_channel )  # (batch, channel)
+    mcs_and_ssim = torch.stack( mcs + [ssim_per_channel], dim = 0 )  # (level, batch, channel)
+    ms_ssim_val = torch.prod( mcs_and_ssim ** weights.view( -1, 1, 1 ), dim = 0 )
 
     if size_average:
         return ms_ssim_val.mean()
     else:
-        return ms_ssim_val.mean(1)
+        return ms_ssim_val.mean( 1 )
 
 
-class SSIM(torch.nn.Module):
+class SSIM( torch.nn.Module ):
     def __init__(
-        self,
-        data_range=255,
-        size_average=True,
-        win_size=11,
-        win_sigma=1.5,
-        channel=3,
-        spatial_dims=2,
-        K=(0.01, 0.03),
-        nonnegative_ssim=False,
+            self,
+            data_range=255,
+            size_average=True,
+            win_size=11,
+            win_sigma=1.5,
+            channel=3,
+            spatial_dims=2,
+            K=(0.01, 0.03),
+            nonnegative_ssim=False,
     ):
         r""" class for ssim
         Args:
@@ -253,9 +251,9 @@ class SSIM(torch.nn.Module):
             nonnegative_ssim (bool, optional): force the ssim response to be nonnegative with relu.
         """
 
-        super(SSIM, self).__init__()
+        super( SSIM, self ).__init__()
         self.win_size = win_size
-        self.win = _fspecial_gauss_1d(win_size, win_sigma).repeat([channel, 1] + [1] * spatial_dims)
+        self.win = _fspecial_gauss_1d( win_size, win_sigma ).repeat( [channel, 1] + [1] * spatial_dims )
         self.size_average = size_average
         self.data_range = data_range
         self.K = K
@@ -265,25 +263,25 @@ class SSIM(torch.nn.Module):
         return ssim(
             X,
             Y,
-            data_range=self.data_range,
-            size_average=self.size_average,
-            win=self.win,
-            K=self.K,
-            nonnegative_ssim=self.nonnegative_ssim,
+            data_range = self.data_range,
+            size_average = self.size_average,
+            win = self.win,
+            K = self.K,
+            nonnegative_ssim = self.nonnegative_ssim,
         )
 
 
-class MS_SSIM(torch.nn.Module):
+class MS_SSIM( torch.nn.Module ):
     def __init__(
-        self,
-        data_range=255,
-        size_average=True,
-        win_size=11,
-        win_sigma=1.5,
-        channel=3,
-        spatial_dims=2,
-        weights=None,
-        K=(0.01, 0.03),
+            self,
+            data_range=255,
+            size_average=True,
+            win_size=11,
+            win_sigma=1.5,
+            channel=3,
+            spatial_dims=2,
+            weights=None,
+            K=(0.01, 0.03),
     ):
         r""" class for ms-ssim
         Args:
@@ -296,9 +294,9 @@ class MS_SSIM(torch.nn.Module):
             K (list or tuple, optional): scalar constants (K1, K2). Try a larger K2 constant (e.g. 0.4) if you get a negative or NaN results.
         """
 
-        super(MS_SSIM, self).__init__()
+        super( MS_SSIM, self ).__init__()
         self.win_size = win_size
-        self.win = _fspecial_gauss_1d(win_size, win_sigma).repeat([channel, 1] + [1] * spatial_dims)
+        self.win = _fspecial_gauss_1d( win_size, win_sigma ).repeat( [channel, 1] + [1] * spatial_dims )
         self.size_average = size_average
         self.data_range = data_range
         self.weights = weights
@@ -308,9 +306,9 @@ class MS_SSIM(torch.nn.Module):
         return ms_ssim(
             X,
             Y,
-            data_range=self.data_range,
-            size_average=self.size_average,
-            win=self.win,
-            weights=self.weights,
-            K=self.K,
+            data_range = self.data_range,
+            size_average = self.size_average,
+            win = self.win,
+            weights = self.weights,
+            K = self.K,
         )
